Refactor: Remove alarm cancellation from TopAppBar
Removed the alarm cancellation logic from the TopAppBar's delete action.
Previously, it was calling `
notificationModel.scheduleNotification` with `onReset = true` to cancel the alarm. This logic is now removed and replaced with a to-do comment indicating the need to implement the cancellation functionality elsewhere.

This change likely moves the responsibility of alarm management to a more appropriate location within the application.