Refactor: Replace File with Uri in DrawingNote
Replaced the usage of `File` with `Uri` for image paths in the `DrawingNote` compos
able. This change likely aims to use URIs for accessing image data, possibly for better compatibility with content providers and modern Android storage practices.

Note that the code is currently commented out and might require further changes to function correctly.