Refactor: Update MediaRepoImpl methods
Renamed `getAllMedias` to `observeAll` and updated it to use `observeAll` from `
dataSource`.
Renamed `addMedia` to `insert` and updated it to use `insert` from `dataSource`.
Renamed `deleteMedia` to `deleteById` and updated it to use `deleteById` from `dataSource`.
Added `observeByUid` method to observe medias by their uid.

These changes likely improve the consistency and clarity of the repository interface and its implementation.