Refactor: Rename `path` to `uri` in Media and MediaMapper
Renamed the `path` property to `uri` in both the `
Media` and `OutMedia` data classes. This change was also reflected in the `MediaMapper`, which now uses `uri` instead of `path` when mapping between the two data types.

This change likely improves the representation of media items, using URIs instead of file paths. Additionally, added documentation to
 `MediaMapper` for better understanding.