feat: Add link delete functionality by uid
The `LinkCard` composable now uses a single `uid` parameter to delete links.
Instead of using
 `noteAndLinkScreenModel` to delete the link's association
and `linkScreenModel` to delete the link itself, the deletion of the link is
now handled through a `sendAction` call to `linkScreenModel` with
`Action.DeleteByUid(uid)`. This simplifies the
 code by using a single unique id
to delete both the link and its relation to a note.

The `noteAndLinkScreenModel` and `noteUid` parameters were also removed.