Refactor: Update AlarmManagerScreenModel to use UpdateById
The `AlarmManagerScreenModel` now uses `UpdateById` use case instead of `Update
` and `UiEvent`.
The `sendUiEvent` method and its related `UiEventHandler` interface are removed.
A new `sendAction` method is added to handle actions using `Action` sealed class.
`withAsync` extension is used to run the `updateById` operation in the background
.
This change simplifies the model and improves data flow.