feat: Add wifi-slash icon to AssistantScreen's input field
Added a `trailingIcon` to the `OutlinedTextField` in the `Assistant
Screen` to display a wifi-slash icon when the device is not connected to the internet. This icon is sourced from `CommonIcons.WIFI_SLASH_ICON`.

The `trailingIcon` provides a visual cue to users when they are unable to send messages due to a lack of connectivity.